requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0
bcrypt==3.2.0
SQLAlchemy==1.4.23
python-jose==3.3.0
Werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
six==1.16.0
pymssql==2.2.7  # 替換 pyodbc